generic_helpline/requirements/common.txt
Django==2.1.1
djangorestframework==3.9.4
requests==2.9.1
celery[redis]==3.1.25
django-celery==3.2.2
loguru==0.5.0
